# openApiSepc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrucksApi",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/Trucks": {
      "get": {
        "tags": [
          "Trucks"
        ],
        "parameters": [
          {
            "name": "NameFilter",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "StatusFilter",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdFilter",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DescriptionFilter",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortFileld",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ascending",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTrucksResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTrucksResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTrucksResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trucks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TruckRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TruckRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TruckRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Trucks/{id}/status": {
      "post": {
        "tags": [
          "Trucks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Trucks/{id}": {
      "put": {
        "tags": [
          "Trucks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTruckRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTruckRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTruckRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Trucks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruckResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Trucks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GetTrucksResponse": {
        "type": "object",
        "properties": {
          "trucks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruckResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TruckRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TruckResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateTruckRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}